@@ -1,8 +1,8 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :app do
-    url "MyString"
-    name "MyString"
-    id_code "MyString"
-    admin nil
+    url {"MyString"}
+    name {"MyString"}
+    id_code {"MyString"}
+    admin {nil}
   end
 end
