@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -12,9 +11,6 @@
 # It's strongly recommended that you check this file into your version control system.
 
 ActiveRecord::Schema.define(version: 20160727061420) do
-
-  # These are extensions that must be enabled in order to support this database
-  enable_extension "plpgsql"
 
   create_table "admins", force: :cascade do |t|
     t.string   "email",                  default: "", null: false
@@ -30,10 +26,9 @@ ActiveRecord::Schema.define(version: 20160727061420) do
     t.datetime "created_at",                          null: false
     t.datetime "updated_at",                          null: false
     t.string   "name"
+    t.index ["email"], name: "index_admins_on_email", unique: true
+    t.index ["reset_password_token"], name: "index_admins_on_reset_password_token", unique: true
   end
-
-  add_index "admins", ["email"], name: "index_admins_on_email", unique: true, using: :btree
-  add_index "admins", ["reset_password_token"], name: "index_admins_on_reset_password_token", unique: true, using: :btree
 
   create_table "apps", force: :cascade do |t|
     t.string   "url"
@@ -42,9 +37,8 @@ ActiveRecord::Schema.define(version: 20160727061420) do
     t.integer  "admin_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.index ["admin_id"], name: "index_apps_on_admin_id"
   end
-
-  add_index "apps", ["admin_id"], name: "index_apps_on_admin_id", using: :btree
 
   create_table "events", force: :cascade do |t|
     t.integer  "app_id"
@@ -52,10 +46,9 @@ ActiveRecord::Schema.define(version: 20160727061420) do
     t.datetime "updated_at", null: false
     t.string   "name"
     t.integer  "user_id"
+    t.index ["app_id"], name: "index_events_on_app_id"
+    t.index ["user_id"], name: "index_events_on_user_id"
   end
-
-  add_index "events", ["app_id"], name: "index_events_on_app_id", using: :btree
-  add_index "events", ["user_id"], name: "index_events_on_user_id", using: :btree
 
   create_table "users", force: :cascade do |t|
     t.string   "name"
@@ -65,12 +58,7 @@ ActiveRecord::Schema.define(version: 20160727061420) do
     t.datetime "created_at",               null: false
     t.datetime "updated_at",               null: false
     t.string   "email"
+    t.index ["app_id"], name: "index_users_on_app_id"
   end
 
-  add_index "users", ["app_id"], name: "index_users_on_app_id", using: :btree
-
-  add_foreign_key "apps", "admins"
-  add_foreign_key "events", "apps"
-  add_foreign_key "events", "users"
-  add_foreign_key "users", "apps"
 end
